backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
alembic==1.15.2
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic-settings==2.8.1
python-dotenv==1.1.0
python-multipart==0.0.20
apscheduler==3.11.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
